requirements.txt
black==23.1.0
    click==8.1.3
    mypy-extensions==1.0.0
    packaging==23.0
    pathspec==0.11.0
    platformdirs==3.1.0
    tomli==2.0.1
    typing-extensions==3.10.0.0
celery==5.2.7
    billiard==3.6.4.0
    # click==8.1.3
    click-didyoumean==0.3.0
        # click==8.1.3
    click-plugins==1.1.1
        # click==8.1.3
    click-repl==0.2.0
        # click==8.1.3
        prompt-toolkit==3.0.38
            wcwidth==0.2.6
        six==1.16.0
    kombu==5.2.4
        amqp==5.1.1
            vine==5.0.0
        # vine==5.0.0
    pytz==2022.7.1
    # vine==5.0.0
django-cors-headers==3.14.0
    Django==4.1.7
        asgiref==3.6.0
        sqlparse==0.4.3
django-fsm==2.8.1
djangorestframework==3.14.0
    # Django==4.1.7
        # asgiref==3.6.0
        # sqlparse==0.4.3
    # pytz==2022.7.1
flake8==6.0.0
    mccabe==0.7.0
    pycodestyle==2.10.0
    pyflakes==3.0.1
gevent==22.10.2
    greenlet==2.0.2
    # setuptools
    zope.event==4.6
        # setuptools
    zope.interface==5.5.2
        # setuptools
gunicorn==20.1.0
    # setuptools
pandas==1.5.3
    numpy==1.24.2
    python-dateutil==2.8.2
        # six==1.16.0
    # pytz==2022.7.1
pipfreeze==2.0.3
pre-commit==3.1.1
    cfgv==3.3.1
    identify==2.5.18
    nodeenv==1.7.0
        # setuptools
    PyYAML==6.0
    virtualenv==20.20.0
        distlib==0.3.6
        filelock==3.9.0
        # platformdirs==3.1.0
psycopg2-binary==2.9.5
python-decouple==3.8
redis==4.5.1
    async-timeout==4.0.2
whitenoise==6.4.0
